V792: implement workaround for duplicating data packets
CAENComm_BLTRead calls CAENVME_FIFOBLTRead with A32_U_BLT as the address
modifier. For unknown reasons it results in data packets duplicated and
some of the header and end of block packets missing. Use
CAENVME_FIFOBLTRead with A32_U_DATA to avoid the problem with the data;
however the board does not assert the bus error when there are no more
data to transfer in this mode.